+++ train_network.py
@@ -1,5 +1,5 @@
 import importlib
-import argparse, ast
+import argparse
 import gc
 import math
 import os
@@ -9,17 +9,22 @@ import time
 import json
 from multiprocessing import Value
 import toml
-import wandb
+
 from tqdm import tqdm
 import torch
 from accelerate.utils import set_seed
 from diffusers import DDPMScheduler
 from library import model_util
-import pickle
+
 import library.train_util as train_util
-from library.train_util import (    DreamBoothDataset,)
+from library.train_util import (
+    DreamBoothDataset,
+)
 import library.config_util as config_util
-from library.config_util import (    ConfigSanitizer,    BlueprintGenerator,)
+from library.config_util import (
+    ConfigSanitizer,
+    BlueprintGenerator,
+)
 import library.huggingface_util as huggingface_util
 import library.custom_train_functions as custom_train_functions
 from library.custom_train_functions import (
@@ -28,18 +33,12 @@ from library.custom_train_functions import (
     prepare_scheduler_for_custom_training,
     scale_v_prediction_loss_like_noise_prediction,
     add_v_prediction_like_loss,
-    apply_gor_loss_precalculated
 )
-from library.group_orthogonalization_normalization import (
-    precalculate_modules_to_check)
-try:
-    from setproctitle import setproctitle
-except (ImportError, ModuleNotFoundError):
-    setproctitle = lambda x: x
-
 def arg_as_list(s):
+    import ast
     v = ast.literal_eval(s)
     return v
+
 class NetworkTrainer:
 
     def __init__(self):
@@ -48,12 +47,9 @@ class NetworkTrainer:
 
     # TODO 他のスクリプトと共通化する
     def generate_step_logs(
-        self, args: argparse.Namespace, current_loss, avr_loss, lr_scheduler, keys_scaled=None, mean_norm=None, maximum_norm=None,
-            task_loss=None, reg_loss=None):
-        logs = {"loss/total current": current_loss,
-                "loss/total average": avr_loss,
-                "loss/task current": task_loss,
-                "loss/reg average": reg_loss, }
+        self, args: argparse.Namespace, current_loss, avr_loss, lr_scheduler, keys_scaled=None, mean_norm=None, maximum_norm=None
+    ):
+        logs = {"loss/current": current_loss, "loss/average": avr_loss}
 
         if keys_scaled is not None:
             logs["max_norm/keys_scaled"] = keys_scaled
@@ -106,8 +102,9 @@ class NetworkTrainer:
     def is_text_encoder_outputs_cached(self, args):
         return False
 
-    def cache_text_encoder_outputs_if_needed(self, args, accelerator, unet, vae, tokenizers, text_encoders,
-                                             data_loader, weight_dtype):
+    def cache_text_encoder_outputs_if_needed(
+        self, args, accelerator, unet, vae, tokenizers, text_encoders, data_loader, weight_dtype
+    ):
         for t_enc in text_encoders:
             t_enc.to(accelerator.device)
 
@@ -142,7 +139,7 @@ class NetworkTrainer:
         tokenizer = self.load_tokenizer(args)
         tokenizers = tokenizer if isinstance(tokenizer, list) else [tokenizer]
 
-        #
+        # データセットを準備する
         if args.dataset_class is None:
             blueprint_generator = BlueprintGenerator(ConfigSanitizer(True, True, False, True))
             if use_user_config:
@@ -214,17 +211,6 @@ class NetworkTrainer:
         accelerator = train_util.prepare_accelerator(args)
         is_main_process = accelerator.is_main_process
 
-        if is_main_process:
-            save_base_folder = os.path.join(args.output_dir)
-            os.makedirs(save_base_folder, exist_ok=True)
-            with open(os.path.join(save_base_folder, 'config.json'), 'w') as f:
-                json.dump(vars(args), f, indent=4)
-
-        if args.log_with == "wandb" :
-            if is_main_process:
-                wandb.init(project=args.wandb_project,)
-                wandb.run.name = args.wandb_run_name
-
         # mixed precisionに対応した型を用意しておき適宜castする
         weight_dtype, save_dtype = train_util.prepare_dtype(args)
         vae_dtype = torch.float32 if args.no_half_vae else weight_dtype
@@ -252,11 +238,14 @@ class NetworkTrainer:
                     multiplier = 1.0
                 else:
                     multiplier = args.base_weights_multiplier[i]
+
                 accelerator.print(f"merging module: {weight_path} with multiplier {multiplier}")
+
                 module, weights_sd = network_module.create_network_from_weights(
                     multiplier, weight_path, vae, text_encoder, unet, for_inference=True
                 )
                 module.merge_to(text_encoder, unet, weights_sd, weight_dtype, accelerator.device if args.lowram else "cpu")
+
             accelerator.print(f"all weights merged: {', '.join(args.base_weights)}")
 
         # 学習を準備する
@@ -289,18 +278,17 @@ class NetworkTrainer:
         if args.dim_from_weights:
             network, _ = network_module.create_network_from_weights(1, args.network_weights, vae, text_encoder, unet, **net_kwargs)
         else:
-            # LyCORIS will work with this...
-            network = network_module.create_network(
-                1.0,
-                args.network_dim,
-                args.network_alpha,
-                vae,
-                text_encoder,
-                unet,
-                block_wise=args.block_wise,
-                neuron_dropout=args.network_dropout,
-                **net_kwargs,
-            )
+            network = network_module.create_network_blockwise(
+                    1.0,
+                    args.network_dim,
+                    args.network_alpha,
+                    vae,
+                    text_encoder,
+                    unet,
+                    block_wise=args.block_wise,
+                    neuron_dropout=args.network_dropout,
+                    **net_kwargs,)
+
         if network is None:
             return
 
@@ -338,15 +326,8 @@ class NetworkTrainer:
                 "Deprecated: use prepare_optimizer_params(text_encoder_lr, unet_lr, learning_rate) instead of prepare_optimizer_params(text_encoder_lr, unet_lr)"
             )
             trainable_params = network.prepare_optimizer_params(args.text_encoder_lr, args.unet_lr)
-        all_params = []
-        for trainable_param in trainable_params:
-            lr = trainable_param["lr"]
-            params = trainable_param["params"]
-            for param in params:
-                param_dict = {"lr": lr, "params": param}
-                all_params.append(param_dict)
-        optimizer_name, optimizer_args, optimizer = train_util.get_optimizer(args, all_params)
-        #optimizer_name, optimizer_args, optimizer = train_util.get_optimizer(args, trainable_params)
+
+        optimizer_name, optimizer_args, optimizer = train_util.get_optimizer(args, trainable_params)
 
         # dataloaderを準備する
         # DataLoaderのプロセス数：0はメインプロセスになる
@@ -693,7 +674,8 @@ class NetworkTrainer:
         progress_bar = tqdm(range(args.max_train_steps), smoothing=0, disable=not accelerator.is_local_main_process, desc="steps")
         global_step = 0
 
-        noise_scheduler = DDPMScheduler(beta_start=0.00085, beta_end=0.012, beta_schedule="scaled_linear", num_train_timesteps=1000, clip_sample=False
+        noise_scheduler = DDPMScheduler(
+            beta_start=0.00085, beta_end=0.012, beta_schedule="scaled_linear", num_train_timesteps=1000, clip_sample=False
         )
         prepare_scheduler_for_custom_training(noise_scheduler, accelerator.device)
         if args.zero_terminal_snr:
@@ -703,8 +685,9 @@ class NetworkTrainer:
             init_kwargs = {}
             if args.log_tracker_config is not None:
                 init_kwargs = toml.load(args.log_tracker_config)
-            accelerator.init_trackers("network_train" if args.log_tracker_name is None else args.log_tracker_name,
-                                      init_kwargs=init_kwargs)
+            accelerator.init_trackers(
+                "network_train" if args.log_tracker_name is None else args.log_tracker_name, init_kwargs=init_kwargs
+            )
 
         loss_list = []
         loss_total = 0.0
@@ -720,18 +703,17 @@ class NetworkTrainer:
         def save_model(ckpt_name, unwrapped_nw, steps, epoch_no, force_sync_upload=False):
             os.makedirs(args.output_dir, exist_ok=True)
             ckpt_file = os.path.join(args.output_dir, ckpt_name)
+
             accelerator.print(f"\nsaving checkpoint: {ckpt_file}")
             metadata["ss_training_finished_at"] = str(time.time())
             metadata["ss_steps"] = str(steps)
             metadata["ss_epoch"] = str(epoch_no)
+
             metadata_to_save = minimum_metadata if args.no_metadata else metadata
             sai_metadata = train_util.get_sai_model_spec(None, args, self.is_sdxl, True, False)
             metadata_to_save.update(sai_metadata)
 
-
-            unwrapped_nw.save_weights(ckpt_file,
-                                      save_dtype,
-                                      metadata_to_save)
+            unwrapped_nw.save_weights(ckpt_file, save_dtype, metadata_to_save)
             if args.huggingface_repo_id is not None:
                 huggingface_util.upload(args, ckpt_file, "/" + ckpt_name, force_sync_upload=force_sync_upload)
 
@@ -741,88 +723,49 @@ class NetworkTrainer:
                 accelerator.print(f"removing old checkpoint: {old_ckpt_file}")
                 os.remove(old_ckpt_file)
 
-        if args.gor_regularization:
-            modules_to_regularize = precalculate_modules_to_check(network.unet_loras, args.gor_name_to_regularize, args.gor_regularize_fc_layers)
-            accelerator.print(f"modules to regularize: {len(modules_to_regularize)}")
-        else:
-            modules_to_regularize = None
-        # print([k for k,v in unet.named_modules() if check_need_to_regularize(v, k, True, ["'up_blocks.*_lora\.up'"])])
         # training loop
-        def merge_conv(lora_down, lora_up):
-            in_rank, in_size, kernel_size, k_ = lora_down.shape
-            out_size, out_rank, _, _ = lora_up.shape
-            assert in_rank == out_rank and kernel_size == k_, f"rank {in_rank} {out_rank} or kernel {kernel_size} {k_} mismatch"
-            merged = lora_up.reshape(out_size, -1) @ lora_down.reshape(in_rank, -1)
-            weight = merged.reshape(out_size, in_size, kernel_size, kernel_size)
-            del lora_up, lora_down
-            return weight
-
-        loras = network.unet_loras + network.text_encoder_loras
-        org_layer_dict = {}
-        for lora in loras:
-            print(f'blockwise lora block name : {lora.lora_name}')
-            if lora.is_linear:
-                lora_name = lora.lora_name
-                org_weight = lora.org_weight
-                org_layer_dict[lora_name] = org_weight
-            else:
-                lora_name = lora.lora_name
-                org_weight = lora.org_weight
-                org_layer_dict[lora_name] = org_weight
-
-        def get_weight(network):
-            loras = network.unet_loras + network.text_encoder_loras
-            layer_dict = {}
-            for lora in loras:
-                if lora.is_linear:
-                    lora_name = lora.lora_name
-                    up_weight = lora.lora_up.weight
-                    down_weight = lora.lora_down.weight
-                    layer_dict[lora_name] = up_weight @ down_weight
-                else:
-                    lora_name = lora.lora_name
-                    lora_up = lora.lora_up.weight
-                    lora_down = lora.lora_down.weight
-                    layer_dict[lora_name] = merge_conv(lora_down, lora_up)
-            return layer_dict
-
-        # training loop
-        gradient_dict = {}
-        task_loss_dict, reg_loss_dict, total_loss_dict = {}, {}, {}
-
         for epoch in range(num_train_epochs):
             accelerator.print(f"\nepoch {epoch+1}/{num_train_epochs}")
             current_epoch.value = epoch + 1
+
             metadata["ss_epoch"] = str(epoch + 1)
+
             network.on_epoch_start(text_encoder, unet)
+
             for step, batch in enumerate(train_dataloader):
                 current_step.value = global_step
                 with accelerator.accumulate(network):
                     on_step_start(text_encoder, unet)
+
                     with torch.no_grad():
                         if "latents" in batch and batch["latents"] is not None:
                             latents = batch["latents"].to(accelerator.device)
                         else:
                             # latentに変換
                             latents = vae.encode(batch["images"].to(dtype=vae_dtype)).latent_dist.sample()
+
                             # NaNが含まれていれば警告を表示し0に置き換える
                             if torch.any(torch.isnan(latents)):
                                 accelerator.print("NaN found in latents, replacing with zeros")
                                 latents = torch.where(torch.isnan(latents), torch.zeros_like(latents), latents)
                         latents = latents * self.vae_scale_factor
                     b_size = latents.shape[0]
+
                     with torch.set_grad_enabled(train_text_encoder):
                         # Get the text embedding for conditioning
                         if args.weighted_captions:
-                            text_encoder_conds = get_weighted_text_embeddings(tokenizer,
+                            text_encoder_conds = get_weighted_text_embeddings(
+                                tokenizer,
                                 text_encoder,
                                 batch["captions"],
                                 accelerator.device,
                                 args.max_token_length // 75 if args.max_token_length else 1,
-                                clip_skip=args.clip_skip,)
+                                clip_skip=args.clip_skip,
+                            )
                         else:
-                            text_encoder_conds = self.get_text_cond(args, accelerator, batch, tokenizers,
-                                                                    text_encoders, weight_dtype)
+                            text_encoder_conds = self.get_text_cond(
+                                args, accelerator, batch, tokenizers, text_encoders, weight_dtype
+                            )
 
                     # Sample noise, sample a random timestep for each image, and add noise to the latents,
                     # with noise offset and/or multires noise if specified
@@ -844,6 +787,7 @@ class NetworkTrainer:
 
                     loss = torch.nn.functional.mse_loss(noise_pred.float(), target.float(), reduction="none")
                     loss = loss.mean([1, 2, 3])
+
                     loss_weights = batch["loss_weights"]  # 各sampleごとのweight
                     loss = loss * loss_weights
 
@@ -853,56 +797,13 @@ class NetworkTrainer:
                         loss = scale_v_prediction_loss_like_noise_prediction(loss, timesteps, noise_scheduler)
                     if args.v_pred_like_loss:
                         loss = add_v_prediction_like_loss(loss, timesteps, noise_scheduler, args.v_pred_like_loss)
-                    # to use this, use --gor_name_to_regularize="lora_up" --gor_regularization=True --gor_regularization_type="inter"
-                    if args.gor_regularization: # required args : gor_num_groups : int, gor_regularization_type: str, gor_name_to_regularize: str, gor_regularize_fc_layers: bool, gor_ortho_decay: float
-                        loss = apply_gor_loss_precalculated(loss, modules_to_regularize, args.gor_num_groups, args.gor_regularization_type, args.gor_ortho_decay)
-                    loss = loss.mean()  # 平均なのでbatch_sizeで割る必要なし
-                    task_loss = loss
-                    task_loss_dict[global_step] = loss.item()
-                    # ----------------------------------------------------------------------------------------------------------------
-                    # reg loss
-                    org_means, org_stds = [], []
-                    new_means, new_stds = [], []
-                    layer_dict = get_weight(network)
-                    for layer in layer_dict.keys():
-                        org_weight = org_layer_dict[layer]
-                        new_weight = layer_dict[layer]
-                        org_mean = torch.mean(org_weight)
-                        new_mean = torch.mean(new_weight)
-                        org_std = torch.std(org_weight)
-                        new_std = torch.std(new_weight)
-                        org_means.append(org_mean)
-                        org_stds.append(org_std)
-                        new_means.append(new_mean)
-                        new_stds.append(new_std)
-                    org_means = torch.stack(org_means)
-                    org_stds = torch.stack(org_stds)
-                    new_means = torch.stack(new_means)
-                    new_stds = torch.stack(new_stds)
-                    reg_loss = torch.mean((org_means - new_means) ** 2) + torch.mean((org_stds - new_stds) ** 2)
-                    reg_loss_dict[global_step] = reg_loss.item()
 
-                    loss = loss * (1-args.reg_loss_weight) + reg_loss * args.reg_loss_weight
-                    # loss = loss + reg_loss * args.reg_loss_weight
-                    total_loss_dict[global_step] = loss.item()
+                    loss = loss.mean()  # 平均なのでbatch_sizeで割る必要なし
+
                     accelerator.backward(loss)
                     if accelerator.sync_gradients and args.max_grad_norm != 0.0:
                         params_to_clip = network.get_trainable_params()
                         accelerator.clip_grad_norm_(params_to_clip, args.max_grad_norm)
-
-                    if is_main_process:
-                        wandb_logs = {}
-                        logs = {}
-                        for (layer_name, param), param_dict in zip(network.named_parameters(), optimizer.param_groups):
-                            wandb_logs[layer_name] = param_dict['params'][0].grad.data.norm(2)
-                            try:
-                                gradient_dict[layer_name].append(param_dict['params'][0].grad.data.norm(2).item())
-                            except:
-                                gradient_dict[layer_name] = []
-                                gradient_dict[layer_name].append(param_dict['params'][0].grad.data.norm(2).item())
-                        if args.log_with == 'wandb':
-                            if is_main_process:
-                                wandb.log(wandb_logs, step=global_step)
 
                     optimizer.step()
                     lr_scheduler.step()
@@ -910,7 +811,8 @@ class NetworkTrainer:
 
                 if args.scale_weight_norms:
                     keys_scaled, mean_norm, maximum_norm = network.apply_max_norm_regularization(
-                        args.scale_weight_norms, accelerator.device)
+                        args.scale_weight_norms, accelerator.device
+                    )
                     max_mean_logs = {"Keys Scaled": keys_scaled, "Average key norm": mean_norm}
                 else:
                     keys_scaled, mean_norm, maximum_norm = None, None, None
@@ -919,15 +821,16 @@ class NetworkTrainer:
                 if accelerator.sync_gradients:
                     progress_bar.update(1)
                     global_step += 1
+
                     self.sample_images(accelerator, args, None, global_step, accelerator.device, vae, tokenizer, text_encoder, unet)
-                    # 指定ステップごとにモデルを保存
                     """
+                    # 指定ステップごとにモデルを保存
                     if args.save_every_n_steps is not None and global_step % args.save_every_n_steps == 0:
                         accelerator.wait_for_everyone()
                         if accelerator.is_main_process:
                             ckpt_name = train_util.get_step_ckpt_name(args, "." + args.save_model_as, global_step)
-                            save_model(ckpt_name,
-                                       accelerator.unwrap_model(network), global_step, epoch)
+                            save_model(ckpt_name, accelerator.unwrap_model(network), global_step, epoch)
+
                             if args.save_state:
                                 train_util.save_and_remove_state_stepwise(args, accelerator, global_step)
 
@@ -936,6 +839,7 @@ class NetworkTrainer:
                                 remove_ckpt_name = train_util.get_step_ckpt_name(args, "." + args.save_model_as, remove_step_no)
                                 remove_model(remove_ckpt_name)
                     """
+
                 current_loss = loss.detach().item()
                 if epoch == 0:
                     loss_list.append(current_loss)
@@ -951,8 +855,7 @@ class NetworkTrainer:
                     progress_bar.set_postfix(**{**max_mean_logs, **logs})
 
                 if args.logging_dir is not None:
-                    logs = self.generate_step_logs(args, current_loss, avr_loss, lr_scheduler, keys_scaled, mean_norm, maximum_norm,
-                                                   )
+                    logs = self.generate_step_logs(args, current_loss, avr_loss, lr_scheduler, keys_scaled, mean_norm, maximum_norm)
                     accelerator.log(logs, step=global_step)
 
                 if global_step >= args.max_train_steps:
@@ -967,24 +870,17 @@ class NetworkTrainer:
             # 指定エポックごとにモデルを保存
             if args.save_every_n_epochs is not None:
                 saving = (epoch + 1) % args.save_every_n_epochs == 0 and (epoch + 1) < num_train_epochs
-
                 if is_main_process and saving:
                     ckpt_name = train_util.get_epoch_ckpt_name(args, "." + args.save_model_as, epoch + 1)
-                    print(f"ckpt_name : {ckpt_name}")
-                    #ckpt_file = os.path.join(args.output_dir, ckpt_name)
-                    #accelerator.save_state(ckpt_file)
-                    unwrap_model = accelerator.unwrap_model(network)
-                    save_model(ckpt_name,
-                               unwrap_model,
-                               #network,
-                               global_step, epoch + 1)
+                    if args.unwrap :
+                        save_model(ckpt_name, accelerator.unwrap_model(network), global_step, epoch + 1)
+                    else :
+                        save_model(ckpt_name, network, global_step, epoch + 1)
 
-                    print("save Lora weights")
                     remove_epoch_no = train_util.get_remove_epoch_no(args, epoch + 1)
                     if remove_epoch_no is not None:
                         remove_ckpt_name = train_util.get_epoch_ckpt_name(args, "." + args.save_model_as, remove_epoch_no)
                         remove_model(remove_ckpt_name)
-                        print("why removing model..?")
 
                     if args.save_state:
                         train_util.save_and_remove_state_on_epoch_end(args, accelerator, epoch + 1)
@@ -1007,40 +903,20 @@ class NetworkTrainer:
         if is_main_process:
             ckpt_name = train_util.get_last_ckpt_name(args, "." + args.save_model_as)
             save_model(ckpt_name, network, global_step, num_train_epochs, force_sync_upload=True)
+
             print("model saved.")
 
-        def save_dictionary (save_base, trg_dict, filename) :
-            save_dir = os.path.join(save_base, filename)
-            with open(save_dir, 'wb') as fw:
-                pickle.dump(trg_dict, fw)
-
-        save_dictionary(gradient_dict, 'gradient_norm.pickle')
-        save_dictionary(task_loss_dict, 'task_loss_dict.pickle')
-        save_dictionary(reg_loss_dict, 'reg_loss_dict.pickle')
-        save_dictionary(total_loss_dict, 'total_loss_dict.pickle')
-
-def add_gor_args(parser: argparse.ArgumentParser)-> None:
-    # required args : gor_num_groups : int, gor_regularization_type: str, gor_name_to_regularize: str, gor_regularize_fc_layers: bool, gor_ortho_decay: float
-    # num_groups 32, ortho_decay 1e-6, str_filters 'up_blocks.*_lora\.up', reg_type 'inter' or 'intra', regularize_fc_layers : True
-    parser.add_argument("--gor_num_groups", type=int, default=32, help="number of groups for group orthogonality regularization")
-    parser.add_argument("--gor_regularization_type", type=str, default=None, help="type of group orthogonality regularization, 'inter' or 'intra'")
-    parser.add_argument("--gor_name_to_regularize", type=str, default='up_blocks.*_lora\.up', help="name of the layer to regularize, e.g. 'up_blocks.*_lora\.up'")
-    parser.add_argument("--gor_regularize_fc_layers", type=bool, default=True, help="whether to regularize fully connected layers")
-    parser.add_argument("--gor_ortho_decay", type=float, default=1e-6, help="decay for group orthogonality regularization")
-    # whether to enable gor_regularization
-    parser.add_argument("--gor_regularization", type=bool, default=False, help="whether to enable group orthogonality regularization")
-
-def add_proctitle_args(parser: argparse.ArgumentParser)-> None:
-    parser.add_argument("--proctitle", type=str, default=None, help="process title")
 
 def setup_parser() -> argparse.ArgumentParser:
     parser = argparse.ArgumentParser()
+
     train_util.add_sd_models_arguments(parser)
     train_util.add_dataset_arguments(parser, True, True, True)
     train_util.add_training_arguments(parser, True)
     train_util.add_optimizer_arguments(parser)
     config_util.add_config_arguments(parser)
     custom_train_functions.add_custom_train_arguments(parser)
+
     parser.add_argument("--no_metadata", action="store_true", help="do not save metadata in output model / メタデータを出力先モデルに保存しない")
     parser.add_argument(
         "--save_model_as",
@@ -1110,23 +986,17 @@ def setup_parser() -> argparse.ArgumentParser:
         action="store_true",
         help="do not use fp16/bf16 VAE in mixed precision (use float VAE) / mixed precisionでも fp16/bf16 VAEを使わずfloat VAEを使う",
     )
-    add_gor_args(parser)
-    add_proctitle_args(parser)
-
     return parser
+
 
 
 if __name__ == "__main__":
     parser = setup_parser()
-    parser.add_argument("--reg_loss_weight", type=float, default=0)
-    parser.add_argument("--wandb_project", type=str)
-    parser.add_argument("--wandb_run_name", type=str)
-    parser.add_argument("--block_wise", type=arg_as_list,
-                        default = '[1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1]')
+    parser.add_argument("--block_wise", type = arg_as_list,
+                        default = [1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,])
+    parser.add_argument("--unwrap", action = 'store_true')
     args = parser.parse_args()
     args = train_util.read_config_from_file(args, parser)
-    # if args.proctitle is not None, set process title
-    if args.proctitle is not None:
-        setproctitle(args.proctitle)
+
     trainer = NetworkTrainer()
     trainer.train(args)
